+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The dates are when development finished, not when the change was actually deploy
 
 Entries in bold are ones important enough that I would have put them in [the old changelog](./CHANGELOG-old.md).
 
+* 2020-02-02: Players are assigned a college when they are generated. [[1]](https://basketball-gm.com/blog/2020/02/colleges/)
+
 * 2020-02-01: In FBGM, touchbacks on punts put the ball at the 20 yard line, not 25.
 
 * 2020-02-01: Added sacks to the league leaders page.